pyproject.toml
[project]
name = "langfuse-summarizer-demo"
version = "0.1.0"
description = "OpenAI text summarization with Langfuse local observability"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langfuse>=2.60.0",
    "openai>=1.50.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
